feat: Price 객체 생성
