+++ src/main/java/kitchenpos/application/MenuService.java
@@ -13,7 +13,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 @Service
 public class MenuService {
@@ -36,12 +35,6 @@ public class MenuService {
 
     @Transactional
     public Menu create(final Menu menu, List<MenuProduct> menuProducts) {
-        final BigDecimal price = menu.getPrice();
-
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException();
-        }
-
         if (!menuGroupRepository.existsById(menu.getMenuGroup().getId())) {
             throw new IllegalArgumentException();
         }
@@ -53,7 +46,7 @@ public class MenuService {
             sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
         }
 
-        if (isGraterThan(price, sum)) {
+        if (isGraterThan(menu.getPrice(), sum)) {
             throw new IllegalArgumentException();
         }
