+++ src/main/java/kitchenpos/domain/menu/Menu.java
@@ -4,6 +4,7 @@ import kitchenpos.domain.BaseEntity;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.Column;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -13,7 +14,9 @@ import java.math.BigDecimal;
 @AttributeOverride(name = "id", column = @Column(name = "MENU_ID"))
 public class Menu extends BaseEntity {
     private String name;
-    private BigDecimal price;
+
+    @Embedded
+    private Price price;
 
     @ManyToOne
     @JoinColumn(name = "MENU_GROUP_ID")
@@ -22,10 +25,20 @@ public class Menu extends BaseEntity {
     public Menu() {
     }
 
-    public Menu(String name, BigDecimal price, MenuGroup menuGroup) {
+    public Menu(String name, Price price, MenuGroup menuGroup) {
         this.name = name;
         this.price = price;
         this.menuGroup = menuGroup;
+    }
+
+    public Menu(String name, BigDecimal price, MenuGroup menuGroup) {
+        this.name = name;
+        this.price = new Price(price);
+        this.menuGroup = menuGroup;
+    }
+
+    public BigDecimal getPrice() {
+        return price.getPrice();
     }
 
     public Long getId() {
@@ -38,10 +51,6 @@ public class Menu extends BaseEntity {
 
     public String getName() {
         return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
     }
 
     public MenuGroup getMenuGroup() {
