+++ src/main/java/kitchenpos/repository/OrderLineItemRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.Orderz;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.Orderz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
